Removing unnecessary assertion
Rationale: Since Swift 2.0 and the new error handling model, a lot of methods (and initializers) that used to include an error pointer and an optional return value now return non-optional values. The String(contentsOfFile:) init is one such method where this is the case, so it makes no sense to check for a nil value on contentsOfStringsFile as it is either guaranteed to return a non-nil value or throw an error.