Use regex for finding translations in strings file
This is done to make matching multiple line comments easier.
Given the new regex approach the comment should be allowed
to span multiple lines already. Tests are missing in this commit.

Previously the string was separated into lines and each line
was matched against a regex which either would result in a
translation line or a comment line. That made it hard to match
multiple line comments.

This change was done to fulfill the request in #44.