Fix build of tests
When building the tests, Xcode complained about not being able to import module “bartycrouch” as “BartyCrouch” in all the imports of the tests.

This seems to be a name clash, though I’m not sure where it’s coming from.
Renaming the Framework and the module seems to fix the issue. This is more of
a workaround than a fix though… :-/