Fixes an issue with renamed Strings files
Given the situation that a user renames their Localizable.strings file but doesn't use
a table name (for whatever reason) the extraction was successful but BartyCrouch
couldn't find the file as it looked for the custom name .strings file, though the extraction
resulted in a /tmpstrings/Localizable.strings file. This adds a fallback to Localizable.strings
for such cases. Table names should still work.